Reworks the minigame card a bit and gates to single completion. Removes old props that are no longer needed
