+++ apps/launchpad/app/components/minigame-card.tsx
@@ -1,22 +1,27 @@
 import { Button, Card, Text } from '@0xintuition/1ui'
 
-import type { Minigame } from '@lib/types/minigame'
 import { useNavigate } from '@remix-run/react'
 
 interface MinigameCardProps {
   className?: string
   onStart: () => void
-  game: Minigame
+  title: string
+  description: string
+  points: number
   hideCTA?: boolean
   isLoading?: boolean
+  resultsLink?: string
 }
 
 export function MinigameCard({
   className,
   onStart,
-  game,
+  title,
+  description,
+  points,
   hideCTA = false,
   isLoading = false,
+  resultsLink,
 }: MinigameCardProps) {
   const navigate = useNavigate()
 
@@ -27,18 +32,17 @@ export function MinigameCard({
       <div className="absolute inset-0 flex flex-col justify-between p-8">
         <div className="space-y-2">
           <Text variant="headline" weight="medium" className="text-foreground">
-            {game.title}
+            {title}
           </Text>
           <Text variant="body" weight="medium" className="text-foreground/70">
-            {game.totalAtoms.toLocaleString()} atoms •{' '}
-            {game.totalUsers.toLocaleString()} users
+            {description}
           </Text>
         </div>
 
         {!hideCTA && (
           <div className="flex flex-col items-center">
             <Button
-              onClick={onStart}
+              onClick={points > 0 ? () => navigate(resultsLink || '') : onStart}
               variant="primary"
               size="lg"
               className="min-w-[200px]"
@@ -46,32 +50,24 @@ export function MinigameCard({
             >
               {isLoading
                 ? 'Loading...'
-                : game.points > 0
-                  ? 'Play Again'
+                : points > 0
+                  ? 'View Results'
                   : 'Earn 200 IQ Points'}
             </Button>
           </div>
         )}
 
-        <div className="flex justify-between items-center">
-          {game.points > 0 ? (
+        <div className="flex justify-end items-center">
+          {points > 0 && (
             <div className="flex items-baseline gap-2">
               <span className="text-xl font-bold bg-gradient-to-r from-[#34C578] to-[#00FF94] bg-clip-text text-transparent">
-                {game.points}
+                {points}
               </span>
               <span className="text-md font-semibold text-muted-foreground">
                 IQ Earned
               </span>
             </div>
-          ) : (
-            <div />
           )}
-          <Button
-            variant="secondary"
-            onClick={() => navigate('/minigames/game-1')}
-          >
-            View Game List
-          </Button>
         </div>
       </div>
     </Card>
